% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flames_import.R
\name{create_config}
\alias{create_config}
\title{Preprocessing functions from \code{FLAMES}}
\usage{
create_config(outdir, type = "sc_3end", ...)
}
\description{
Preprocessing functions from \code{FLAMES}
}
