% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flames_import.R
\name{create_sce_from_dir}
\alias{create_sce_from_dir}
\title{Create \code{SingleCellExperiment} object from \code{FLAMES} output folder}
\usage{
create_sce_from_dir(outdir, annotation, quantification = "FLAMES")
}
\description{
Create \code{SingleCellExperiment} object from \code{FLAMES} output folder
}
