% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gtf.R
\name{filter_gtf}
\alias{filter_gtf}
\title{Filter gtf file with valid transcript_id}
\usage{
filter_gtf(input_file, output_file)
}
\arguments{
\item{input_file}{The raw gtf file}

\item{output_file}{The path to the filtered gtf file}
}
\description{
Filter gtf file with valid transcript_id
}
