% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Create a heatmap visualizing cell-level transcript usage of a gene between 2 gRNA groups}
\usage{
plot_heatmap(
  final_sce,
  c1,
  c2,
  gene,
  annot,
  min_entry = 5,
  n_clusters = 3,
  random_seed = 0
)
}
\arguments{
\item{final_sce}{The \code{SingleCellExperiment} object with cell labels}

\item{c1}{The first gRNA group}

\item{c2}{The second gRNA group}

\item{gene}{The gene of interest}

\item{annot}{The annotation file}

\item{min_entry}{The threshold for the filtering step}

\item{n_clusters}{The number of clusters for K-means}

\item{random_seed}{The random seed}
}
\value{
A heatmap showing the cell-level transcript usage of the gene between
2 groups, with a column showing K-means clusters, and a column showing the
total molecule counts of the gene in each cell
}
\description{
Create a heatmap visualizing cell-level transcript usage of a gene between 2 gRNA groups
}
