% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_gene.R
\name{quantify_gene}
\alias{quantify_gene}
\title{Function from \code{FLAMES}}
\usage{
quantify_gene(
  annotation,
  outdir,
  infq,
  n_process,
  pipeline = "sc_single_sample",
  samples = NULL,
  random_seed = 2024
)
}
\description{
Function from \code{FLAMES}
}
