% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_DTU_analysis.R
\name{sc_DTU_analysis}
\alias{sc_DTU_analysis}
\title{Performing DTU analysis between 2 gRNA groups}
\usage{
sc_DTU_analysis(
  sce,
  gRNA_dir,
  method,
  gRNA1,
  gRNA2,
  min_count = 0.001,
  filter_by_entry = FALSE,
  min_entry = 5
)
}
\arguments{
\item{sce}{The \code{SingleCellExperiment} object generated from the \code{FLAMES} preprocessing step}

\item{gRNA_dir}{The directory containg gRNA files}

\item{method}{The chosen statistical method, containing "chisq" (Chi-squared),
"multi" (multinomial), and "dir-multi" (Dirichlet-Multinomial)}

\item{gRNA1}{The first gRNA of interest}

\item{gRNA2}{The second gRNA of interest}

\item{min_count}{The threshold used in the \code{FLAMES} filtering approach}

\item{filter_by_entry}{The boolean value to change to \code{CRISPRana} filtering approach}

\item{min_entry}{The threshold used in the \code{CRISPRana} filtrting approach}
}
\value{
a \code{tibble}
}
\description{
Performing DTU analysis between 2 gRNA groups
}
