% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flames_import.R
\name{sc_long_pipeline}
\alias{sc_long_pipeline}
\title{The pipeline for the preprocessing step from \code{FLAMES}}
\usage{
sc_long_pipeline(
  annotation,
  fastq,
  genome_bam = NULL,
  outdir,
  genome_fa,
  minimap2 = NULL,
  k8 = NULL,
  barcodes_file = NULL,
  expect_cell_number = NULL,
  config_file = NULL
)
}
\description{
The pipeline for the preprocessing step from \code{FLAMES}
}
